requirements.txt
numpy
Flask
tensorflow-cpu
matplotlib
gunicorn
